requirements.txt
anthropic>=0.80.0
openai>=1.0.0
pydantic>=2.0.0
chromadb>=0.4.0
pyyaml>=6.0
python-dotenv>=1.0.0
faker>=20.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
